backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-keycloak>=3.5.0
keycloak>=0.0.0

# Task Queue
celery>=5.3.6
redis>=5.0.1

# AI/ML
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
